+++ src/random.h
@@ -37,7 +37,7 @@ public:
     double normal(double mean=0, double sd=1);
     void poisson(std::vector<int>&, double mean=1);
     int poisson(double mean=1);
-    double binomial (int n, double p);
+    size_t binomial (int n, double p);
 ///@}
 
 ///@name Additional distribution
@@ -46,7 +46,7 @@ public:
   \param n_frequence (std::vector<double>&): frequencies that need to be modified by the function
   \return new frequencies, randomly generated using a combination of \ref binomial
   */
-    std::vector<double> multinomial (int N, std::vector<double> n_frequence);
+    std::vector<double> multinomial (size_t N, std::vector<double> n_frequence);
 /*! @name Auxiliary function
   This takes a vector of indices and re-orders it randomly.
  */
